Fixed posible integer overflow in posix events

+++ Scheduler/Include/Platform/Posix/MTEvent.h
@@ -185,7 +185,7 @@ namespace MT
 			struct timeval tv;
 			gettimeofday(&tv, NULL);
 
-			uint64_t nanoseconds = ((uint64_t) tv.tv_sec) * 1000 * 1000 * 1000 + milliseconds * 1000 * 1000 + ((uint64_t) tv.tv_usec) * 1000;
+			uint64_t nanoseconds = ((uint64_t) tv.tv_sec) * 1000 * 1000 * 1000 + (uint64_t)milliseconds * 1000 * 1000 + ((uint64_t) tv.tv_usec) * 1000;
 
 			ts.tv_sec = nanoseconds / 1000 / 1000 / 1000;
 			ts.tv_nsec = (nanoseconds - ((uint64_t) ts.tv_sec) * 1000 * 1000 * 1000);
